Fix things up and move the preprocessor into rend3
