+++ rend3-routine/shaders/src/depth.wgsl
@@ -1,4 +1,4 @@
-{{include "structures.wgsl"}}
+{{include "rend3-routine/structures.wgsl"}}
 
 struct VertexInput {
     @location(0) position: vec3<f32>,
