+++ rend3-routine/shaders/src/skybox.wgsl
@@ -1,4 +1,4 @@
-{{include "structures.wgsl"}}
+{{include "rend3-routine/structures.wgsl"}}
 
 struct VertexOutput {
     @builtin(position) pos: vec4<f32>,
@@ -24,8 +24,8 @@ fn fs_main(output: VertexOutput) -> @location(0) vec4<f32> {
     // We use the near plane as depth here, as if we used the far plane, it would all NaN out. Doesn't _really_ matter,
     // but 1.0 is a nice round number and results in a depth of 0.1 with my near plane. Good 'nuf.
     let clip = vec4<f32>(output.clip_position, 1.0, 1.0);
-    let world = uniforms.inv_origin_view_proj * clip;
-    let world = world.xyz / world.w;
+    let world_undiv = uniforms.inv_origin_view_proj * clip;
+    let world = world_undiv.xyz / world_undiv.w;
     let world_dir = normalize(world);
 
     let background = textureSample(skybox, primary_sampler, world_dir).rgb;
