+++ rend3-routine/src/skybox.rs
@@ -9,7 +9,7 @@ use rend3::{
     },
     types::{SampleCount, TextureHandle},
     util::bind_merge::{BindGroupBuilder, BindGroupLayoutBuilder},
-    Renderer,
+    Renderer, ShaderConfig, ShaderPreProcessor,
 };
 use wgpu::{
     BindGroup, BindGroupLayout, BindingType, Color, ColorTargetState, ColorWrites, CompareFunction, DepthBiasState,
@@ -18,10 +18,7 @@ use wgpu::{
     ShaderStages, StencilState, TextureFormat, TextureSampleType, TextureViewDimension, VertexState,
 };
 
-use crate::{
-    common::WholeFrameInterfaces,
-    shaders::{ShaderConfig, ShaderPreProcessor},
-};
+use crate::common::WholeFrameInterfaces;
 
 struct StoredSkybox {
     bg: Option<BindGroup>,
@@ -155,7 +152,7 @@ impl SkyboxPipelines {
         let skybox_sm = renderer.device.create_shader_module(ShaderModuleDescriptor {
             label: Some("skybox vert"),
             source: ShaderSource::Wgsl(Cow::Owned(
-                spp.render_shader("skybox.wgsl", &ShaderConfig::default()).unwrap(),
+                spp.render_shader("rend3-routine/skybox.wgsl", &ShaderConfig::default()).unwrap(),
             )),
         });
