+++ rend3/src/managers/material.rs
@@ -295,7 +295,9 @@ impl MaterialManager {
             BindingType::Buffer {
                 ty: BufferBindingType::Storage { read_only: true },
                 has_dynamic_offset: false,
-                min_binding_size: NonZeroU64::new((round_up_pot(M::TEXTURE_COUNT * 4, 16) + round_up_pot(M::DATA_SIZE, 16)) as _),
+                min_binding_size: NonZeroU64::new(
+                    (round_up_pot(M::TEXTURE_COUNT * 4, 16) + round_up_pot(M::DATA_SIZE, 16)) as _,
+                ),
             },
             None,
         );
